[ui] Export GhostDaggy svg
[INTERNAL_BRANCH=dish/fe-772-deployment-picker]
